Create README.md
This code uses the Selenium WebDriver to automate the process of searching for a one-way flight from Brussels to New York on the Expedia website. The ChromeDriver is set up in the setup() method and the website is loaded in the loading() method, both of which are called before the main @Test method. In the @Test method, the code navigates through the website, entering in the departure and destination cities, selecting the date and number of passengers, and searching for the flight. It uses xpaths to locate elements on the webpage and interacts with them (clicking, sending keys, selecting from dropdown etc) .It also uses a while loop to check if the next month is April 2023 and if it is, it breaks out of the loop, otherwise it continues to the next month.